refactor: extract App utilities to lib modules
